+++ odra-casper/shared/src/consts.rs
@@ -46,11 +46,11 @@ pub const ALLOW_KEY_OVERRIDE_ARG: &str = "odra_cfg_allow_key_override";
 /// The arg name for the contract upgradeability setting.
 pub const IS_UPGRADABLE_ARG: &str = "odra_cfg_is_upgradable";
 
+/// Constructor name argument.
+pub const CONSTRUCTOR_NAME_ARG: &str = "odra_cfg_constructor";
+
 /// The state key name.
 pub const STATE_KEY: &str = "state";
 
 /// Constructor group name.
 pub const CONSTRUCTOR_GROUP_NAME: &str = "constructor_group";
-
-/// Constructor name argument.
-pub const CONSTRUCTOR_NAME_ARG: &str = "constructor";